CRUD for posts, replies, reposts, likes and following has been added and updated

+++ backend/controllers/UsersController/DeleteUserFollow.js
@@ -1,27 +1,32 @@
 const asyncHandler = require('express-async-handler');
 const { Users, Following } = require('../../models/userModel');
+const { ObjectId } = require('mongodb');
 
 const DeleteUserFollow = asyncHandler(async (req, res) => {
-    const { userID, userFollowingID } = req.body;
+    const { followID, userID } = req.query;
+
+    console.log(followID, userID)
 
     const userExists = await Users.findById(userID);
 
     if (userExists) {
-        const userFollowingExist = await Following.findOne({
-            $and: [
-                { followingID: userFollowingID },
-                { followerID: userID }
-            ]
-        });
+        const userFollowingExist = await Following.findById(followID);
 
         if (userFollowingExist) {
-            await userFollowingExist.deleteOne()
-            const unfollowID = {
-                followingID: userFollowingExist.followingID,
-                followerID: userFollowingExist.followerID
+
+            if (userFollowingExist.followerID.toString() === userID) {
+                await userFollowingExist.deleteOne()
+                const unfollowData = {
+                    followID: userFollowingExist._id,
+                    followingID: userFollowingExist.followingID
+                }
+
+                res.status(202).json(unfollowData)
+            } else {
+
+                res.status(401).json({ message: "Unauthorized" })
             }
 
-            res.status(202).json({ unfollowID })
         } else {
 
             res.status(204).json({ message: "You did not follow this user" })
